% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_identified_params.R
\name{get_identified_params}
\alias{get_identified_params}
\alias{get_identified_params_vi}
\title{Extract identified parameters from MV probit}
\usage{
get_identified_params(mv_probit_fit, M, P, K)

get_identified_params_vi(model_fit, M, P, K, n_samps)
}
\arguments{
\item{mv_probit_fit}{the fit from the multivariate probit model}

\item{M}{dimension of the covariance matrix; number of responses.}

\item{P}{number of covariates; excluding intercept}

\item{K}{number of factors used}

\item{model_fit}{the fit from the multivariate probit model}

\item{n_samps}{number of samples to draw}

\item{n_iter}{number of iterations the sampler was run for}

\item{burn_in}{number of iterations to drop from the sampler run}

\item{M}{dimension of the covariance matrix; number of responses.}

\item{P}{number of covariates; excluding intercept}

\item{K}{number of factors used}
}
\description{
Extract identified parameters from MV probit

Extract identified parameters from MV probit svi fit
}
