% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_hs_gibbs}
\alias{lm_hs_gibbs}
\title{Univariate normal linear regression with a Horseshoe prior using a Gibbs
sampler.}
\usage{
lm_hs_gibbs(y, X, verbose = TRUE, n_iter = 10000L, a_tau = 0.1, b_tau = 0.1)
}
\arguments{
\item{y}{Vector or responses (N by 1)}

\item{X}{Matrix of predictors (N by P)}

\item{verbose}{True of False. Do you want to print messages along the way?}

\item{n_iter}{Number of iterations to run the algorithm for}

\item{a_tau}{Prior shape parameter for the likelihood precision.}

\item{b_tau}{Prior rate parameter for the likelihood precision.}
}
\description{
Univariate normal linear regression with a Horseshoe prior using a Gibbs
sampler.
}
