% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_hs_gibbs_gammas}
\alias{lm_hs_gibbs_gammas}
\title{update the latent variables for the double exponential prior}
\usage{
lm_hs_gibbs_gammas(tau, lambda, nus, b, P, gammas)
}
\description{
update the latent variables for the double exponential prior
}
