% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_hs_gibbs_tau}
\alias{lm_hs_gibbs_tau}
\title{updates the precision parameter tau in a bayesian hs regression}
\usage{
lm_hs_gibbs_tau(ehat, b, lambda, gammas, a_tau, b_tau, N, P, tau)
}
\description{
updates the precision parameter tau in a bayesian hs regression
}
