% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_lasso_gibbs_b}
\alias{lm_lasso_gibbs_b}
\title{function to update the coefficients in a linear regression}
\usage{
lm_lasso_gibbs_b(X, y_tilde, gammas, tau, P, b)
}
\arguments{
\item{X}{N by P matrix of covariates}
}
\description{
function to update the coefficients in a linear regression
}
