% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_lasso_gibbs_lambda2}
\alias{lm_lasso_gibbs_lambda2}
\title{updates the prior precision parameter of b, lambda, in a bayesian lasso
regression}
\usage{
lm_lasso_gibbs_lambda2(gammas, a_lambda, b_lambda, P, lambda2)
}
\description{
updates the prior precision parameter of b, lambda, in a bayesian lasso
regression
}
