% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_ridge_gibbs_b}
\alias{lm_ridge_gibbs_b}
\title{function to update the coefficients in a linear regression}
\usage{
lm_ridge_gibbs_b(X, y, lambda, tau, b, P)
}
\arguments{
\item{X}{N by P matrix of covariates}
}
\description{
function to update the coefficients in a linear regression
}
