% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_ridge_gibbs_lambda}
\alias{lm_ridge_gibbs_lambda}
\title{updates the prior precision parameter of b, lambda, in a bayesian ridge
regression}
\usage{
lm_ridge_gibbs_lambda(b, tau, lambda, a_lambda, b_lambda, P)
}
\description{
updates the prior precision parameter of b, lambda, in a bayesian ridge
regression
}
