% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_ridge_gibbs_tau}
\alias{lm_ridge_gibbs_tau}
\title{updates the precision parameter tau in a bayesian ridge regression}
\usage{
lm_ridge_gibbs_tau(ehat, b, lambda, tau, a_tau, b_tau, N, P)
}
\description{
updates the precision parameter tau in a bayesian ridge regression
}
