% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_ridge_svi}
\alias{lm_ridge_svi}
\title{Univariate normal linear regression with a ridge (normal) prior using the
SVI algorithm.}
\usage{
lm_ridge_svi(
  y,
  X,
  n_iter = 15000L,
  verbose = TRUE,
  a_tau = 0.1,
  b_tau = 0.1,
  a_lambda = 0.1,
  b_lambda = 0.1,
  type = 0L,
  batch_size = 42L,
  const_rhot = 0.01,
  omega = 15,
  kappa = 0.6
)
}
\arguments{
\item{y}{Vector or responses (N by 1)}

\item{X}{Matrix of predictors (N by P)}

\item{n_iter}{Number of iterations to run the algorithm for (default =
5000).}

\item{verbose}{True or False. Do you want to print messages along the way?}

\item{a_tau}{Prior shape parameter for the likelihood precision.}

\item{b_tau}{Prior rate parameter for the likelihood precision.}

\item{a_lambda}{Prior shape parameter for the coefficient precision
(shrinkage) term.}

\item{b_lambda}{Prior rate parameter for the coefficient precision
(shrinkage) term.}

\item{type}{Correlation structure of the regression coefficients. Use 0 for
full correlation and 1 for independece assumption.}

\item{batch_size}{Size of the subsamples used to update the parameters.}

\item{omega}{Delay for the stepsize (\eqn{\omega}) for the gradient step.
Interacts with \eqn{\kappa} via the formula \eqn{\rho_{t} = (t +
\omega)^{-\kappa}}. This parameter has to be greater than or equal to zero.}

\item{kappa}{Forgetting rate for the step size iterations; \eqn{\kappa \in
(0.5, 1)}}

\item{cost_rhot}{Used to set a constant step size in the gradient descent
algorithm. If this parameter is greater than zero, it overrides the step
size iterations calculated using kappa and omega.}
}
\description{
Univariate normal linear regression with a ridge (normal) prior using the
SVI algorithm.
}
