% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_lm_uninf_vi.R
\name{mv_lm_uninf_svi}
\alias{mv_lm_uninf_svi}
\title{Multivariate linear regression with a factor model}
\usage{
mv_lm_uninf_svi(
  Y,
  X,
  K = 2,
  n_iter,
  verbose = TRUE,
  a_tau = 0.1,
  b_tau = 0.1,
  batch_size = 42,
  const_rhot = 0.01,
  omega = 15,
  kappa = 0.6
)
}
\arguments{
\item{Y}{matrix of responses}

\item{X}{matrix of predictors to control for}

\item{K}{number of factors in the factor model}

\item{n_iter}{number of iterations to run the Gibbs sampler}

\item{verbose}{True or False. Print status of the sampler.}

\item{batch_size}{Size of the subsamples used to update the parameters.}

\item{omega}{Delay for the stepsize (\eqn{\omega}) for the gradient step.
Interacts with \eqn{\kappa} via the formula \eqn{\rho_{t} = (t +
\omega)^{-\kappa}}. This parameter has to be greater than or equal to zero.}

\item{kappa}{Forgetting rate for the step size iterations; \eqn{\kappa \in
(0.5, 1)}}

\item{cost_rhot}{Used to set a constant step size in the gradient descent
algorithm. If this parameter is greater than zero, it overrides the step
size iterations calculated using kappa and omega.}
}
\description{
Run a Gibbs sampler for the multivariate probit model. Uses the
multiplicative gamma process for the factor structure
}
