% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_probit_uninf_vi.R
\name{mv_probit_uninf_cavi}
\alias{mv_probit_uninf_cavi}
\title{multivariate probit with factor model - variational}
\usage{
mv_probit_uninf_cavi(Y, X, K, n_iter = 100, n_samps = 1000,
  verbose = TRUE, rel_tol = 1e-05)
}
\arguments{
\item{Y}{integer matrix for the responses}

\item{X}{matrix of predictors to control for}

\item{K}{number of factors in the factor model}

\item{n_iter}{number of iterations to run the algorithm}

\item{n_samps}{number of samples to draw from the variational distribution}

\item{verbose}{print status of the sampler}

\item{rel_tol}{relative tolerance at which to end the algorithm}
}
\description{
multivariate probit with factor model - variational
}
