% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_probit_uninf_vi.R
\name{mv_probit_uninf_svi}
\alias{mv_probit_uninf_svi}
\title{multivariate probit with factor model - variational}
\usage{
mv_probit_uninf_svi(
  Y,
  X,
  K,
  n_iter = 10000,
  n_samps = 1000,
  verbose = TRUE,
  batch_size = 10,
  const_rhot = 0.001,
  omega = 15,
  kappa = 0.6
)
}
\arguments{
\item{Y}{integer matrix for the responses}

\item{X}{matrix of predictors to control for}

\item{K}{number of factors in the factor model}

\item{n_iter}{number of iterations to run the algorithm}

\item{verbose}{print status of the sampler}

\item{batch_size}{Size of the subsamples used to update the parameters.}

\item{omega}{Delay for the stepsize (\eqn{\omega}) for the gradient step.
Interacts with \eqn{\kappa} via the formula \eqn{\rho_{t} = (t +
\omega)^{-\kappa}}. This parameter has to be greater than or equal to zero.}

\item{kappa}{Forgetting rate for the step size iterations; \eqn{\kappa \in
(0.5, 1)}}

\item{samps}{number samples from the variational distribution}

\item{cost_rhot}{Used to set a constant step size in the gradient descent
algorithm. If this parameter is greater than zero, it overrides the step
size iterations calculated using kappa and omega.}
}
\value{
A list with samples from the variational distribution for the
  intercept term (\code{b0_mat}), regression coefficients (\code{B_array}),
  and upper triangular elements of the correlation matrix
  (\code{cor_mat_upper_tri})
}
\description{
multivariate probit with factor model - variational
}
