% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funs.R
\name{predict_lm_vi}
\alias{predict_lm_vi}
\title{Predictions from Normal VI regression model fits}
\usage{
predict_lm_vi(fit, X_test)
}
\arguments{
\item{fit}{the model fit}

\item{X_test}{matrix of predictor values for which you want to predict the
response}
}
\description{
Predictions from Normal VI regression model fits
}
