% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_lm_uninf_gibbs.R
\name{predict_mv_lm}
\alias{predict_mv_lm}
\title{predict from a multivariate linear model}
\usage{
predict_mv_lm(mvlm_fit, X_test)
}
\arguments{
\item{mvlm_fit}{fit from a mv_lm}
}
\value{
an array with predictions (i, N, M)
}
\description{
predict from a multivariate linear model
}
