% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_lm_uninf_vi.R
\name{predict_mv_lm_vi}
\alias{predict_mv_lm_vi}
\title{generate predictions for a multivariate fit from VI algorithms}
\usage{
predict_mv_lm_vi(fit, X_test, n_samps)
}
\arguments{
\item{fit}{model fit}

\item{X_test}{matrix of predictor values at which to predict the response}

\item{n_samps}{number of samples to draw from the variational predictive
distribution}
}
\value{
Array of draws from the predictive distribution, along with the means
  of the distributions, and their lower (2.5) and upper (97.5) bounds of the
  credible intervals.
}
\description{
generate predictions for a multivariate fit from VI algorithms
}
