% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funs.R
\name{predict_probit_gibbs}
\alias{predict_probit_gibbs}
\title{Predictions from probit Gibbs regression model fits}
\usage{
predict_probit_gibbs(model_fit, X_test, seq_to_keep)
}
\arguments{
\item{X_test}{matrix of predictor values for which you want to predict the
response}

\item{fit}{the model fit}
}
\description{
Predictions from probit Gibbs regression model fits
}
