% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funs.R
\name{predict_probit_vi}
\alias{predict_probit_vi}
\title{Predictions from probit VI regression model fits}
\usage{
predict_probit_vi(model_fit, X_test)
}
\arguments{
\item{X_test}{matrix of predictor values for which you want to predict the
response}

\item{fit}{the model fit}
}
\description{
Predictions from probit VI regression model fits
}
