% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probit_hs_gibbs}
\alias{probit_hs_gibbs}
\title{Univariate probit linear regression with a Horseshoe prior using a
Gibbs sampler.}
\usage{
probit_hs_gibbs(y, X, verbose = TRUE, n_iter = 10000L)
}
\arguments{
\item{y}{Vector or responses (N by 1)}

\item{X}{Matrix of predictors (N by P)}

\item{verbose}{True of False. Do you want to print messages along the way?
(shrinkage) term.}

\item{n_iter}{Number of iterations to run the algorithm for}
}
\description{
Univariate probit linear regression with a Horseshoe prior using a
Gibbs sampler.
}
