% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probit_lasso_gibbs}
\alias{probit_lasso_gibbs}
\title{Univariate probit linear regression with a LASSO prior using a
Gibbs sampler.}
\usage{
probit_lasso_gibbs(y, X, verbose = TRUE, n_iter = 10000L,
  a_lambda2 = 0.1, b_lambda2 = 0.1)
}
\arguments{
\item{y}{Vector or responses (N by 1)}

\item{X}{Matrix of predictors (N by P)}

\item{verbose}{True or False. Do you want to print messages along the way?}

\item{n_iter}{Number of iterations to run the algorithm for}

\item{a_lambda2}{Prior shape parameter for the coefficient precision
(shrinkage) term.}

\item{b_lambda2}{Prior rate parameter for the coefficient precision
(shrinkage) term.}
}
\description{
Univariate probit linear regression with a LASSO prior using a
Gibbs sampler.
}
