% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{probit_ridge_gibbs}
\alias{probit_ridge_gibbs}
\title{Univariate probit linear regression with a ridge (normal) prior using a
Gibbs sampler.}
\usage{
probit_ridge_gibbs(
  y,
  X,
  verbose = TRUE,
  n_iter = 10000L,
  a_lambda = 0.01,
  b_lambda = 0.01
)
}
\arguments{
\item{y}{Vector or responses (N by 1)}

\item{X}{Matrix of predictors (N by P)}

\item{verbose}{True of False. Do you want to print messages along the way?}

\item{n_iter}{Number of iterations to run the algorithm for}

\item{a_lambda}{Prior shape parameter for the coefficient precision
(shrinkage) term.}

\item{b_lambda}{Prior rate parameter for the coefficient precision
(shrinkage) term.}
}
\description{
Univariate probit linear regression with a ridge (normal) prior using a
Gibbs sampler.
}
