% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_lm_uninf_vi.R
\docType{data}
\name{sim_dat}
\alias{sim_dat}
\title{Multivariate linear regression with a factor model}
\format{An object of class \code{list} of length 2.}
\usage{
sim_dat
}
\arguments{
\item{Y}{matrix of responses}

\item{X}{matrix of predictors to control for}

\item{K}{number of factors in the factor model}

\item{n_iter}{number of iterations to run the Gibbs sampler}

\item{verbose}{True or False. Print status of the sampler.}
}
\value{
A list with samples for the intercept term (\code{b0_mat}),
  regression coefficients (\code{B_array}), and upper triangular elements of
  the correlation matrix (\code{cor_mat_upper_tri})
}
\description{
Run a Gibbs sampler for the multivariate probit model. Uses the
multiplicative gamma process for the factor structure
}
\keyword{datasets}
