% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credible_intervals.R
\name{summary_vi}
\alias{summary_vi}
\title{Credible intervals from VI model fits}
\usage{
summary_vi(fit, level = 0.95, coef_names = NULL)
}
\arguments{
\item{fit}{the model fit}

\item{level}{level of the credible interval. Default is 95\%.}

\item{coef_names}{vector of names for the predictors. Use `colnames(X)` to
generate.}
}
\description{
Credible intervals from VI model fits
}
