% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_lm_uninf_vi.R
\name{mv_lm_uninf_cavi}
\alias{mv_lm_uninf_cavi}
\title{Multivariate linear regression with a factor model - CAVI}
\usage{
mv_lm_uninf_cavi(
  Y,
  X,
  K = 2,
  n_iter = 1000,
  rel_tol = 1e-04,
  verbose = TRUE,
  a_tau = 0.1,
  b_tau = 0.1
)
}
\arguments{
\item{Y}{matrix of responses}

\item{X}{matrix of predictors to control for}

\item{K}{number of factors in the factor model}

\item{n_iter}{number of iterations to run the Gibbs sampler}

\item{verbose}{True or False. Print status of the sampler.}
}
\value{
A list with samples for the intercept term (\code{b0_mat}),
  regression coefficients (\code{B_array}), and upper triangular elements of
  the correlation matrix (\code{cor_mat_upper_tri})
}
\description{
Estimate the parameters in a multivariate linear model with the CAVI
algorithm. The model uses a multiplicative gamma process for the factor
structure
}
